changes to output of bad imp_markers
