+++ src/_version.py
@@ -1,1 +1,1 @@
-__version__ = "0.6.8"
+__version__ = "0.6.9"
